Cart fix

+++ LaravelProject/app/Http/Controllers/OrderController.php
@@ -13,17 +13,22 @@ class OrderController extends Controller
    function order (Request $req){
        $user = JWTAuth::user()->id;
        $carts = Cart::where('user_id', $user)->get(); 
-       foreach ($carts as $cart){
-          $order = new Order();
-          $order->user_id = $cart['user_id'];
-          $order->product_id = $cart['product_id'];
-          $order->quantity = $cart['quantity'];
-          $order->address = $req->input('address');
-          $order->total = $cart['sub_total'];
-          $order->payment_method = $req->input('payment_method');
-          $order->save();
-          DB::table('cart')->where('user_id', $user)->delete();
-       }            
+       if (empty($carts) || $carts == null){
+           return response()->json(['message' =>'your cart is empty'], 500);
+       }
+       else{
+        foreach ($carts as $cart){
+            $order = new Order();
+            $order->user_id = $cart['user_id'];
+            $order->product_id = $cart['product_id'];
+            $order->quantity = $cart['quantity'];
+            $order->address = $req->input('address');
+            $order->total = $cart['sub_total'];
+            $order->payment_method = $req->input('payment_method');
+            $order->save();
+            DB::table('cart')->where('user_id', $user)->delete();
+        }    
+     } return;       
    }
